requirements.txt
mysql-connector
flask
flask_mail
pandas
lxml
beautifulsoup4